requirements.txt
certifi==2018.4.16
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==2.1.7
django-allauth==0.44.0
django-crispy-forms==1.7.2
django-datetime-widget2==0.9.5
django-mailer==2.1
django-money==0.14.2
django-widget-tweaks==1.4.8
gunicorn==20.1.0
idna==2.7
lockfile==0.12.2
mailchimp3==3.0.4
mysqlclient==1.3.13
oauthlib==3.1.0
py-moneyed==0.7.0
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.4
python-magic==0.4.15
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.23
